week03: defined entities

+++ cinema/src/main/java/com/nagarro/af24/cinema/model/Genre.java
@@ -1,6 +1,11 @@
 package com.nagarro.af24.cinema.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -15,7 +20,7 @@ public class Genre {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "genre_id")
-    private int genreId;
+    private Long genreId;
 
     private String type;
 }
